addplant and plantprofile separate routes work
