+++ routes/plant/addplant.js
@@ -1,6 +1,10 @@
 const express = require('express');
 const router = express.Router();
 
+// Import schemas
+const User = require('../../schemas/user');
+const Plant = require('../../schemas/plant');
+
 // Authenticate user
 const authenticateUser = async (req, res, next) => {
     const accessToken = req.header("Authorization");
@@ -21,3 +25,31 @@ const authenticateUser = async (req, res, next) => {
       })
     }
   };
+
+  // Add plant to garden
+router.post("/", authenticateUser);
+router.post("/", async (req, res) => {
+  try {
+  const { plantname, species } = req.body;
+  const accessToken = req.header("Authorization");
+  const user = await User.findOne({accessToken: accessToken});
+  const plants = await new Plant({
+    plantname: plantname,
+    species: species,
+    user: user._id
+  }).save()
+  res.status(200).json({
+    success: true,
+    response: {
+      plants: plants
+    }
+})
+} catch (e) {
+  res.status(500).json({
+    success: false,
+    response: e
+  })
+}
+});
+
+module.exports = router;
