+++ routes/plant/deleteplant.js
@@ -1,6 +1,10 @@
 const express = require('express');
 const router = express.Router();
 
+// Import schemas
+const User = require('../../schemas/user');
+const Plant = require('../../schemas/plant');
+
 // Authenticate user
 const authenticateUser = async (req, res, next) => {
     const accessToken = req.header("Authorization");
