+++ routes/user/register.js
@@ -1,2 +1,5 @@
 const express = require('express');
 const router = express.Router();
+
+// Import schemas
+const User = require('../../schemas/user');
