+++ routes/user/userprofile.js
@@ -1,6 +1,9 @@
 const express = require('express');
 const router = express.Router();
 
+// Import schemas
+const User = require('../../schemas/user');
+
 // Authenticate user
 const authenticateUser = async (req, res, next) => {
     const accessToken = req.header("Authorization");
